# docs/openapi.json
{
  "openapi": "3.0.1",
  "paths": {
...
    "/api/projects/{projectId}/items": {
      "get": {
        "operationId": "getItems",
        "parameters": [
          {
            "description": "The project",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectItems"
                }
              }
            },
            "description": "default response"
          }
        },
        "summary": "Returns the open issues and discussions of all repository settings of a project, with what the import would do with each",
        "tags": [
          "Import"
        ]
      }
    },
...
    "/api/projects/{projectId}/repositories/{name}/import": {
      "post": {
        "operationId": "importRepository",
        "parameters": [
          {
            "description": "The project that gets the work items",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The name of the repository setting",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "True to report only what the import would do",
            "in": "query",
            "name": "dryRun",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          }
        },
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            },
            "description": "default response"
          }
        },
        "summary": "Creates work items from the open issues and discussions of a configured repository",
        "tags": [
          "Import"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ImportEntry": {
        "description": "The outcome of the import for one GitHub item",
        "properties": {
          "assignees": {
            "description": "The GitHub logins the item is assigned to",
            "items": {
              "description": "The GitHub logins the item is assigned to",
              "type": "string"
            },
            "type": "array"
          },
          "githubType": {
            "description": "The issue type, or the category of a discussion",
            "type": "string"
          },
          "kind": {
            "description": "Issue or discussion",
            "enum": [
              "ISSUE",
              "DISCUSSION"
            ],
            "type": "string"
          },
          "labels": {
            "description": "The labels of the item",
            "items": {
              "description": "The labels of the item",
              "type": "string"
            },
            "type": "array"
          },
          "message": {
            "description": "The reason of a failure",
            "type": "string"
          },
          "number": {
            "description": "The number of the item in its repository",
            "format": "int64",
            "type": "integer"
          },
          "repository": {
            "description": "The repository as owner/name",
            "type": "string"
          },
          "setting": {
            "description": "The name of the repository setting",
            "type": "string"
          },
          "status": {
            "description": "What the import did, or would do",
            "enum": [
              "NEW",
              "CREATED",
              "EXISTS",
              "SKIPPED",
              "FAILED"
            ],
            "type": "string"
          },
          "title": {
            "description": "The title of the item on GitHub",
            "type": "string"
          },
          "url": {
            "description": "The URL of the item on GitHub",
            "type": "string"
          },
          "workItemAssignees": {
            "description": "The names of the users the work item is assigned to",
            "items": {
              "description": "The names of the users the work item is assigned to",
              "type": "string"
            },
            "type": "array"
          },
          "workItemId": {
            "description": "The ID of the work item, when one exists or was created",
            "type": "string"
          },
          "workItemStatus": {
            "description": "The status of the work item",
            "type": "string"
          },
          "workItemType": {
            "description": "The ID of the work item type: of the work item, or the one the import would create",
            "type": "string"
          },
          "workItemTypeName": {
            "description": "The name of that work item type",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ImportRequest": {
        "description": "Which items of a repository to import",
        "properties": {
          "urls": {
            "description": "The GitHub URLs to import. Leave it out to import every open item.",
            "items": {
              "description": "The GitHub URLs to import. Leave it out to import every open item.",
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "ImportResult": {
        "description": "The outcome of the import of one repository",
        "properties": {
          "dryRun": {
            "description": "Whether the import only reported what it would do",
            "type": "boolean"
          },
          "entries": {
            "description": "One entry per GitHub item",
            "items": {
              "$ref": "#/components/schemas/ImportEntry"
            },
            "type": "array"
          },
          "readAt": {
            "description": "When the oldest list of the repository was read from GitHub, as an ISO-8601 instant. Lists serve from a cache for five minutes.",
            "type": "string"
          },
          "repository": {
            "description": "The repository as owner/name",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ProjectItems": {
        "description": "The open GitHub items of all repository settings of a project",
        "properties": {
          "entries": {
            "description": "The open issues and discussions, with what the import would do with each",
            "items": {
              "$ref": "#/components/schemas/ImportEntry"
            },
            "type": "array"
          },
          "repositories": {
            "description": "One state per repository setting",
            "items": {
              "$ref": "#/components/schemas/RepositoryState"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "RepositoryState": {
        "description": "How the reading of one repository setting went",
        "properties": {
          "error": {
            "description": "Why the repository could not be read",
            "type": "string"
          },
          "readAt": {
            "description": "When the oldest list of the repository was read from GitHub, as an ISO-8601 instant",
            "type": "string"
          },
          "repository": {
            "description": "The repository as owner/name",
            "type": "string"
          },
          "setting": {
            "description": "The name of the repository setting",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  }
}